Time: 15 ms (19.12%), Space: 26.2 MB (49.23%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int smallestDivisor(vector<int>& nums, int threshold) {
-        long long maxi=*max_element(nums.begin(),nums.end());
-        long long low=1,high=maxi;
-        long long ans=INT_MAX;
+        int maxi=*max_element(nums.begin(),nums.end());
+        int low=1,high=maxi;
+        int ans=INT_MAX;
         while(low<=high)
         {
-            long long mid=(low+high)/2;
-            long long a=0;
+            int mid=(low+high)/2;
+            int a=0;
             for(int i=0;i<nums.size();i++)
             {
                 a+=ceil((double)nums[i]/mid);
